third_party/blink/web_tests/external/wpt/html/cross-origin-opener-policy/iframe-popup-same-origin-to-same-origin.https-expected.txt

This is a testharness.js-based test.
PASS same-origin with SAME_ORIGIN iframe opening popup a SAME_ORIGIN with COOP: same-origin
PASS same-origin with SAME_SITE iframe opening popup a SAME_ORIGIN with COOP: same-origin
PASS same-origin with CROSS_ORIGIN iframe opening popup a SAME_ORIGIN with COOP: same-origin
FAIL same-origin with SAME_ORIGIN iframe opening popup a SAME_SITE with COOP: same-origin assert_equals: opener expected false but got true
PASS same-origin with SAME_SITE iframe opening popup a SAME_SITE with COOP: same-origin
PASS same-origin with CROSS_ORIGIN iframe opening popup a SAME_SITE with COOP: same-origin
FAIL same-origin with SAME_ORIGIN iframe opening popup a CROSS_ORIGIN with COOP: same-origin assert_equals: opener expected false but got true
PASS same-origin with SAME_SITE iframe opening popup a CROSS_ORIGIN with COOP: same-origin
PASS same-origin with CROSS_ORIGIN iframe opening popup a CROSS_ORIGIN with COOP: same-origin
Harness: the test ran to completion.
